@@ -20,22 +20,16 @@
         }
     ],
     "require": {
-        "php": "^7.3 || ^8.0",
-        "codeigniter/coding-standard": "^1.1",
-        "fakerphp/faker": "^1.9",
-        "mikey179/vfsstream": "^1.6",
-        "nexusphp/cs-config": "^3.1",
-        "nexusphp/tachycardia": "^1.0",
-        "phpstan/extension-installer": "^1.1",
-        "phpstan/phpstan": "^1.0",
-        "phpstan/phpstan-deprecation-rules": "^1.0",
-        "phpstan/phpstan-phpunit": "^1.0",
-        "phpunit/phpunit": "^9.3",
-        "roave/security-advisories": "dev-latest",
-        "tatter/patches": "^2.0"
+        "php": "^7.4 || ^8.0",
+        "codeigniter4/devkit": "^1.0"
     },
     "require-dev": {
-        "codeigniter4/codeigniter4": "dev-develop"
+        "codeigniter4/framework": "^4.1"
+    },
+    "config": {
+        "allow-plugins": {
+            "phpstan/extension-installer": true
+        }
     },
     "autoload": {
         "psr-4": {
@@ -58,6 +52,9 @@
     ],
     "minimum-stability": "dev",
     "prefer-stable": true,
+    "bin": [
+        "src/retool"
+    ],
     "scripts": {
         "analyze": "phpstan analyze",
         "ci": [
@@ -65,13 +62,14 @@
             "@deduplicate",
             "@analyze",
             "@test",
-            "@mutate",
             "@inspect",
+            "rector process",
             "@style"
         ],
         "deduplicate": "phpcpd app/ src/",
         "inspect": "deptrac analyze --cache-file=build/deptrac.cache",
         "mutate": "infection --threads=2 --skip-initial-tests --coverage=build/phpunit",
+        "retool": "retool",
         "style": "php-cs-fixer fix --verbose --ansi --using-cache=no",
         "test": "phpunit"
     }
