@@ -1,21 +1,15 @@
 <?php
 
-/**
- * This file is part of Tatter Tools.
- *
- * (c) 2021 Tatter Software
- *
- * For the full copyright and license information, please view
- * the LICENSE file that was distributed with this source code.
- */
-
 use CodeIgniter\CodingStandard\CodeIgniter4;
 use Nexus\CsConfig\Factory;
 use PhpCsFixer\Finder;
 
 $finder = Finder::create()
     ->files()
-    ->in(__DIR__)
+    ->in([
+        __DIR__ . '/src/',
+        __DIR__ . '/tests/',
+    ])
     ->exclude('build')
     ->append([__FILE__]);
 
@@ -26,9 +20,4 @@ $options = [
     'cacheFile' => 'build/.php-cs-fixer.cache',
 ];
 
-return Factory::create(new CodeIgniter4(), $overrides, $options)->forLibrary(
-    'Tatter Tools',
-    'Tatter Software',
-    '',
-    2021
-);
+return Factory::create(new CodeIgniter4(), $overrides, $options)->forProjects();
